Edicion y ajuste de los archivos enlasados con el seeder productos
