--- app/Http/Controllers/UsuariosController.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class UsuariosController extends Controller
-{
-    //
-    public function index(){
-        return view('/auth/usuarios');
-}
-}
