+++ resources/views/auth/administrador/editar-productos.blade.php
