+++ resources/views/auth/bebidas-alcoholicas.blade.php
@@ -1,6 +1,5 @@
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -8,6 +7,44 @@
     <link href="{{ asset('asset/Productos.css') }}" rel="stylesheet">
     <title>Document</title>
 </head>
+
+
+<body>
+
+<div class="toolbar">
+        <div class="menu" onclick="toggleDropdown()">
+            <div class="dropdown" id="dropdownMenu" style="display: none;">
+                <!-- Elementos del menú -->
+                <a href="/bebidas-alcoholicas">Bebidas Alcoholicas</a>
+                <a href="/hamburguesas">Hamburguesas</a>
+                <a href="/heladeria">Heladeria</a>
+                <a href="/comidas-rapidas">Comidas Rapidas</a>
+                <a href="/bebidas">Bebidas</a>
+                <a href="/combos">Combos</a>
+            </div>
+        </div>
+
+        <div class="botonera">
+            <a>
+                <div class="boton_toolbar1"></div>
+            </a>
+            <a href="/dashboard">
+                <div class="boton_toolbar2"></div>
+            </a>
+            <a href="/bolsa">
+                <div class="boton_toolbar3"></div>
+            </a>
+            <a href="/lista-pedidos">
+                <div class="boton_toolbar4"></div>
+            </a>
+
+            <!--nombre de usuario registrado (perfil)  -->
+            <a>
+                <div class="usuario">usuario</div>
+            </a>
+
+        </div>
+    </div>
 
 <div class="ContainerProductos">
 
@@ -27,8 +64,17 @@
             <h1>{{ $bebida->nombre }}</h1>
             <p>{{ $bebida->Descripción }}</p>
             <div class="botones">
-                <h5>{{ $bebida->Stock }}</< /h5>
-                    <div class="BotonPedir">Pedir</div>
+                <h5>{{ $bebida->Stock }}</h5>
+                <form action="/procesar_pedido" method="POST">
+                    @csrf <!-- Agrega esto si estás utilizando Blade para evitar el error CSRF -->
+                    <div class="botones">
+                        <label for="cantidad">
+                            <h5>Cantidad</h5>
+                        </label>
+                        <input type="number" id="cantidad" name="cantidad" min="1" value="1">
+                    </div>
+                    <button type="submit" id="botonPedir1">Pedir</button>
+                </form>
             </div>
         </div>
     </div>
@@ -36,7 +82,7 @@
 
 </div>
 
-<script src="{{ asset('asset/js/bebidas-alcoholicas.js') }}"></script>
+<script src="{{ asset('asset/js/Productos.js') }}"></script>
 </body>
 
 </html>
