+++ resources/views/auth/heladeria.blade.php
@@ -45,9 +45,10 @@
     <div class="Titulo">Helados</div>
 
     <div class="ContainerProductos">
-        @foreach(helados as helado)
+        @foreach($helados as $helado)
         <div class="Cuadro">
             <div class="img1">
+            <img src="{{ $helado->image}}" alt="Descripción de la imagen">
                 <div class="Calificador" id="calificacion1">
                     <span class="estrella" onclick="calificar(1, 1)">&#9734;</span>
                     <span class="estrella" onclick="calificar(1, 2)">&#9734;</span>
@@ -95,112 +96,9 @@
                 </form>
             </div>
         </div>
+        @endforeach
+    </div>
 
-
-
-
-        <div class="Cuadro">
-            <div class="img1">
-                <div class="Calificador" id="calificacion2">
-                    <span class="estrella" onclick="calificar(2, 1)">&#9734</span>
-                    <span class="estrella" onclick="calificar(2, 2)">&#9734</span>
-                    <span class="estrella" onclick="calificar(2, 3)">&#9734</span>
-                    <span class="estrella" onclick="calificar(2, 4)">&#9734</span>
-                    <span class="estrella" onclick="calificar(2, 5)">&#9734</span>
-                </div>
-            </div>
-            <div class="Detalle">
-                <h1>{{ $helado->nombre }}</h1>
-                <p>{{ $helado->Descripción }}</p>
-
-                <div class="Sabores">
-                    <h3>Sabores:</h3>
-                    <form id="saboresForm">
-                        <label class="Sabor" id="sabor1">
-                            <input type="checkbox" name="sabor" value="Chocolate" class="SaborCheckbox-button">
-                            <div class="SaborImagen"></div>
-                            <span class="texto-descriptivo">Chocolate</span>
-                        </label>
-                        <label class="Sabor" id="sabor2">
-                            <input type="checkbox" name="sabor" value="Vainilla" class="SaborCheckbox-button">
-                            <div class="SaborImagen"></div>
-                            <span class="texto-descriptivo">Vainilla</span>
-                        </label>
-                        <label class="Sabor" id="sabor3">
-                            <input type="checkbox" name="sabor" value="Fresa" class="SaborCheckbox-button">
-                            <div class="SaborImagen"></div>
-                            <span class="texto-descriptivo">Fresa</span>
-                        </label>
-                        <label class="Sabor" id="sabor4">
-                            <input type="checkbox" name="sabor" value="Chicle" class="SaborCheckbox-button">
-                            <div class="SaborImagen"></div>
-                            <span class="texto-descriptivo">Chicle</span>
-                        </label>
-                </div>
-
-                <div class="botones">
-                    <label for="cantidad">
-                        <h4>Cantidad</h4>
-                    </label>
-                    <input type="number" id="cantidad" name="cantidad" min="1" value="1">
-                </div>
-                <button type="submit" id="botonPedir2">Pedir</button>
-                </form>
-            </div>
-        </div>
-
-
-
-        <div class="Cuadro">
-            <div class="img1">
-
-                <div class="Calificador" id="calificacion3">
-                    <span class="estrella" onclick="calificar(3, 1)">&#9734;</span>
-                    <span class="estrella" onclick="calificar(3, 2)">&#9734;</span>
-                    <span class="estrella" onclick="calificar(3, 3)">&#9734;</span>
-                    <span class="estrella" onclick="calificar(3, 4)">&#9734;</span>
-                    <span class="estrella" onclick="calificar(3, 5)">&#9734;</span>
-                </div>
-            </div>
-            <div class="Detalle">
-                <h1>{{ $helado->nombre }}</h1>
-                <p>{{ $helado->Descripción }}</p>
-
-                <div class="Sabores">
-                    <h3>Sabores:</h3>
-                    <form id="saboresForm">
-                        <label class="Sabor" id="sabor1">
-                            <input type="checkbox" name="sabor" value="Chocolate" class="SaborCheckbox-button">
-                            <div class="SaborImagen"></div>
-                            <span class="texto-descriptivo">Chocolate</span>
-                        </label>
-                        <label class="Sabor" id="sabor2">
-                            <input type="checkbox" name="sabor" value="Vainilla" class="SaborCheckbox-button">
-                            <div class="SaborImagen"></div>
-                            <span class="texto-descriptivo">Vainilla</span>
-                        </label>
-                        <label class="Sabor" id="sabor3">
-                            <input type="checkbox" name="sabor" value="Fresa" class="SaborCheckbox-button">
-                            <div class="SaborImagen"></div>
-                            <span class="texto-descriptivo">Fresa</span>
-                        </label>
-                        <label class="Sabor" id="sabor4">
-                            <input type="checkbox" name="sabor" value="Chicle" class="SaborCheckbox-button">
-                            <div class="SaborImagen"></div>
-                            <span class="texto-descriptivo">Chicle</span>
-                        </label>
-                </div>
-
-                <div class="botones">
-                    <label for="cantidad">
-                        <h4>Cantidad</h4>
-                    </label>
-                    <input type="number" id="cantidad" name="cantidad" min="1" value="1">
-                </div>
-                <button type="submit" id="botonPedir3">Pedir</button>
-                </form>
-            </div>
-        </div>
-
-
-        <script src="{{ asset('asset/js/Productos.js') }}"></script>
+    <script src="{{ asset('asset/js/Productos.js') }}"></script>
+</body>
+</html>
