--- resources/views/auth/lista-pedidos.blade.php
@@ -1,43 +1,0 @@
-<link href="{{ asset('asset/lista-pedidos.css') }}" rel="stylesheet">
-
-<div class="ContenedorPedidos">
-    <div class="DescripcionBreve">
-        <div class="Numerador">1</div>
-        <div class="Datos">
-            <h4>Momento:</h4>
-            <a>Productos:.......................................</a><br>
-            <a>Cliente:.......................................</a>
-        </div>
-        <div class="BotonRevisar">Revisar</div>
-    </div>
-
-    <div class="DescripcionBreve">
-        <div class="Numerador">1</div>
-        <div class="Datos">
-            <h4>Momento:</h4>
-            <a>Productos:.......................................</a><br>
-            <a>Cliente:.......................................</a>
-        </div>
-        <div class="BotonRevisar">Revisar</div>
-    </div>
-
-    <div class="DescripcionBreve">
-        <div class="Numerador">1</div>
-        <div class="Datos">
-            <h4>Momento:</h4>
-            <a>Productos:.......................................</a><br>
-            <a>Cliente:.......................................</a>
-        </div>
-        <div class="BotonRevisar">Revisar</div>
-    </div>
-
-    <div class="DescripcionBreve">
-        <div class="Numerador">1</div>
-        <div class="Datos">
-            <h4>Momento:</h4>
-            <a>Productos:.......................................</a><br>
-            <a>Cliente:.......................................</a>
-        </div>
-        <div class="BotonRevisar">Revisar</div>
-    </div>
-</div>
