+++ routes/web.php
@@ -10,8 +10,10 @@ use App\Http\Controllers\HeladeriaController;
 use App\Http\Controllers\ListaPedidosController;
 use App\Http\Controllers\PedidosController;
 use App\Http\Controllers\AdministradorController;
-use App\Http\Controllers\UsuariosController;
+use App\Http\Controllers\EditarProductoController;
+use App\Http\Controllers\NuevosUsuariosController;
 use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\RolesController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -52,11 +54,15 @@ Route::get('heladeria', [HeladeriaController::class, 'index']);
 Route::get('comidas-rapidas', [ComidasRapidasController::class, 'index']);
 Route::get('bebidas', [BebidasController::class, 'index']);
 Route::get('combos', [CombosController::class, 'index']);
-Route::get('pedidos', [PedidosController::class, 'index']);
 Route::get('bolsa', [BolsaController::class, 'index']);
 Route::get('lista-pedidos', [ListaPedidosController::class, 'index']);
+Route::get('pedidos', [PedidosController::class, 'index']);
+Route::get('editar-productos', [EditarProductoController::class, 'index']);
+Route::get('nuevos_usuarios', [NuevosUsuariosController::class, 'index']);
+Route::get('roles', [RolesController::class, 'index']);
 Route::get('admin', [AdministradorController::class, 'index']);
-Route::get('usuarios', [UsuariosController::class, 'index']);
+
+
